Refactoring views
